refactor: extract task manager
